animGrp back to uchar. Added static CGameCharacter::GetTrueAnimationGroup(), it will be called in CharacterAnimation packet handler and transform crazy high action numbers into proper anim groups according to files.

+++ OrionUO/Managers/AnimationManager.h
@@ -271,7 +271,7 @@ public:
 	@param [__in] group Группа анимации
 	@return true в случае успеха
 	*/
-	bool AnimationExists(const ushort &graphic, ushort group);
+	bool AnimationExists(const ushort &graphic, uchar group);
 
 	ANIMATION_DIMENSIONS GetAnimationDimensions(uchar frameIndex, const ushort &id, const uchar &dir, const uchar &animGroup, const bool &isCorpse);
